+++ browser/src/Services/Commands/GlobalCommands.ts
@@ -9,17 +9,13 @@ import { remote } from "electron"
 
 import * as Oni from "oni-api"
 
-import {
-    getAllErrorsForFile,
-    getInstance as getDiagnosticsInstance,
-} from "./../../Services/Diagnostics"
+import { gotoNextError, gotoPreviousError } from "./../../Services/Diagnostics/navigateErrors"
 import { EditorManager } from "./../../Services/EditorManager"
 import { MenuManager } from "./../../Services/Menu"
 import { showAboutMessage } from "./../../Services/Metadata"
 import { multiProcess } from "./../../Services/MultiProcess"
 import { Tasks } from "./../../Services/Tasks"
 import { windowManager } from "./../../Services/WindowManager"
-import { isInRange } from "./../../Utility"
 
 // import * as UI from "./../UI/index"
 
@@ -49,78 +45,6 @@ export const activate = (
     const popupMenuNext = popupMenuCommand(() => menuManager.nextMenuItem())
     const popupMenuPrevious = popupMenuCommand(() => menuManager.previousMenuItem())
     const popupMenuSelect = popupMenuCommand(() => menuManager.selectMenuItem())
-
-    const gotoNextError = async () => {
-        const errors = getDiagnosticsInstance().getErrors()
-        const activeBuffer = editorManager.activeEditor.activeBuffer
-        const currentFileErrors = getAllErrorsForFile(activeBuffer.filePath, errors)
-        const currentPosition = activeBuffer.cursor
-
-        if (!currentFileErrors || currentFileErrors.length === 0) {
-            return
-        }
-
-        for (const error of currentFileErrors) {
-            if (isInRange(currentPosition.line, currentPosition.column, error.range)) {
-                continue
-            }
-
-            const currentLine = (await activeBuffer.getLines(currentPosition.line))[0]
-            if (
-                currentPosition.line === error.range.start.line &&
-                currentLine.length <= error.range.start.character
-            ) {
-                continue
-            }
-
-            if (
-                error.range.start.line > currentPosition.line ||
-                (error.range.start.line === currentPosition.line &&
-                    error.range.start.character > currentPosition.column)
-            ) {
-                await activeBuffer.setCursorPosition(
-                    error.range.start.line,
-                    error.range.start.character,
-                )
-                return
-            }
-        }
-
-        activeBuffer.setCursorPosition(
-            currentFileErrors[0].range.start.line,
-            currentFileErrors[0].range.start.character,
-        )
-    }
-
-    const gotoPreviousError = async () => {
-        const errors = getDiagnosticsInstance().getErrors()
-        const activeBuffer = editorManager.activeEditor.activeBuffer
-        const currentFileErrors = getAllErrorsForFile(activeBuffer.filePath, errors)
-        const currentPosition = activeBuffer.cursor
-
-        if (!currentFileErrors || currentFileErrors.length === 0) {
-            return
-        }
-
-        let lastError = currentFileErrors[currentFileErrors.length - 1]
-        for (const error of currentFileErrors) {
-            if (
-                isInRange(currentPosition.line, currentPosition.column, error.range) ||
-                error.range.start.line > currentPosition.line ||
-                (error.range.start.line === currentPosition.line &&
-                    error.range.start.character > currentPosition.column)
-            ) {
-                await activeBuffer.setCursorPosition(
-                    lastError.range.start.line,
-                    lastError.range.start.character,
-                )
-                return
-            }
-            lastError = error
-        }
-
-        activeBuffer.setCursorPosition(lastError.range.start.line, lastError.range.start.character)
-    }
 
     const commands = [
         new CallbackCommand("editor.executeVimCommand", null, null, (message: string) => {
